+++ CHANGELOG.md
@@ -63,6 +63,8 @@
 ### Removed
 - `Into<Value>`/`From<T> for Value` implementations, use `IntoValue`.
 - `Default` implementations for `Value`, `RClass`, and `ExceptionClass`.
+- `Error` can no longer be directly constructed, must use `Error::new` or
+  `Error::from`/`Exception::into`.
 
 ### Fixed